backend/pyproject.toml
[project]
name = "three-kingdoms-strategy"
version = "0.1.0"
description = "Three Kingdoms Strategy Alliance Manager Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn[standard]>=0.37.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "supabase>=2.21.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in argument defaults (FastAPI pattern)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
